requirements.txt
pyyaml
bs4
requests
bibtexparser
